# [llvm] Use explicit null checks with LLVM.