@imhameed 2019-09-20
> 
> 
> Failed test `gc-altstack.exe` on `Linux x64 FullAOT+LLVM` looks related

Yeah. And the 'devirtualization' fullaot test is failing. I'm debugging these now.