@vargaz 2019-09-20
- Will the extra llvm pass get rid of all the explicit null checks ?
- There is no code to add the make.implicit metadata.