@EgorBo 2019-09-21
```csharp
[MethodImpl(MethodImplOptions.NoInlining)]
static int Foo(string str)
{
    return str.Length;
}
```
Diff: https://www.diffchecker.com/65XrwcrY (implicit null checks on the right)
llvm-jit.