@imhameed 2019-09-22
> 
> 
>     * Will the extra llvm pass get rid of all the explicit null checks ?
> 
>     * There is no code to add the make.implicit metadata.
* No, not all of them.
* This was added in: https://github.com/mono/mono/pull/16668/files#diff-042576030c385dd63e0a2a1a1033b6b8R217