@SamMonoRT 2019-10-10
just fyi - we can already see some significant benchmark improvements, (some minor regressions too, which I believe are ok to submit) : https://microsoft-my.sharepoint.com/:x:/p/imhameed/EZ3MevtMiFREimv_b4KzzccBRWYdhXUcK-KdMEi0jWsU_Q?e=Uxs8aD