@lewurm 2019-10-11
Code size was raised as a concern. Before this PR:
```console
$ find mcs -name 'mscorlib.dll.so' | xargs ls -l
-rwxrwxr-x 1 lewurm lewurm 11999736 Oct 11 05:04 mcs/class/lib/build-linux/mscorlib.dll.so
-rwxrwxr-x 1 lewurm lewurm 16567056 Oct 11 05:05 mcs/class/lib/net_4_x-linux/mscorlib.dll.so
```

with this PR:
```console
$ find mcs -name 'mscorlib.dll.so' | xargs ls -l
-rwxrwxr-x 1 lewurm lewurm 12843656 Oct 11 04:57 mcs/class/lib/build-linux/mscorlib.dll.so
-rwxrwxr-x 1 lewurm lewurm 17459760 Oct 11 04:59 mcs/class/lib/net_4_x-linux/mscorlib.dll.so
```

(built with `./autogen.sh --enable-llvm --with-runtime-preset=aot_llvm` which is actually broken... https://github.com/mono/mono/issues/17284 )

So it's a 5% increase, which would be okay if it has clear perf benefits (there might be some low-hanging fruit to reduce that, see below). In terms of performance, size increase is bad on hot paths because it can trash the instruction cache. That might what have happened on `scimark-mm` where this change causes a 10% slowdown.

What I'm confused about too is with the example from @EgorBo:
Before:
```
_Simple_Foo_string:
    1960:   8b 47 10    movl    16(%rdi), %eax
    1963:   c3  retq
```

After:
```
_Simple_Foo_string:
    1960:   50  pushq   %rax
    1961:   8b 47 10    movl    16(%rdi), %eax
    1964:   59  popq    %rcx
    1965:   c3  retq
    1966:   bf 1c 01 00 00  movl    $284, %edi
    196b:   e8 94 02 00 00  callq   660 <plt__jit_icall_llvm_throw_corlib_exception_abs_trampoline>
```
I assume the last two instructions are left for other purposes, but in our case the `SIGSEGV` handler of mono will handle a `null` deref, e.g. calling `Simple.foo (null)`:
```
(gdb) b Simple_Foo_string
Function "Simple_Foo_string" not defined.
Make breakpoint pending on future shared library load? (y or [n]) y
Breakpoint 1 (Simple_Foo_string) pending.
(gdb) r
Starting program: /home/lewurm/work/mono/mono/mini/mono-sgen Simple.exe
[Thread debugging using libthread_db enabled]
Using host libthread_db library "/lib/x86_64-linux-gnu/libthread_db.so.1".
[New Thread 0x7ffff5bff700 (LWP 103382)]
[New Thread 0x7ffff3499700 (LWP 103383)]

Thread 1 "mono-sgen" hit Breakpoint 1, 0x00007ffff3096c10 in Simple_Foo_string () from /home/lewurm/work/mono/Simple.exe.so
(gdb) x/3i $pc
=> 0x7ffff3096c10 <Simple_Foo_string>:  push   %rax
   0x7ffff3096c11 <Simple_Foo_string+1>:        mov    0x10(%rdi),%eax
   0x7ffff3096c14 <Simple_Foo_string+4>:        pop    %rcx
(gdb) si
0x00007ffff3096c11 in Simple_Foo_string () from /home/lewurm/work/mono/Simple.exe.so
(gdb)

Thread 1 "mono-sgen" received signal SIGSEGV, Segmentation fault.
0x00007ffff3096c11 in Simple_Foo_string () from /home/lewurm/work/mono/Simple.exe.so
(gdb) si
mono_sigsegv_signal_handler (_dummy=11, _info=0x7ffff7fe74b0, context=0x7ffff7fe7380) at mini-runtime.c:3407
3407            mono_sigsegv_signal_handler_debug (MONO_SIG_HANDLER_PARAMS_DEBUG);
```

It would be nice to get rid of those left over instructions. I couldn't figure out how though 😞 might not be that easy, given `llc` is taking care of that.

Another example:
```csharp
        [MethodImpl(MethodImplOptions.NoInlining)]
        static int Bar(string str1, string str2)
        {
                    return str1.Length + str2.Length;
        }
```
The generated code:
```
0000000000000c20 <Simple_Bar_string_string>:
 c20:   50                      push   %rax
 c21:   48 85 ff                test   %rdi,%rdi
 c24:   74 08                   je     c2e <Simple_Bar_string_string+0xe>
 c26:   8b 46 10                mov    0x10(%rsi),%eax
 c29:   03 47 10                add    0x10(%rdi),%eax
 c2c:   59                      pop    %rcx
 c2d:   c3                      retq
 c2e:   bf 29 01 00 00          mov    $0x129,%edi
 c33:   e8 6c 02 00 00          callq  ea4 <plt__jit_icall_llvm_throw_corlib_exception_abs_trampoline>
```
which is again not something I would have expected: Why is it able to transform one load to implicit but keep the other explicit?

Overall I'm not so sure anymore about this change. Sorry if this raises more questions than answers 😕